requirements.txt
pyserial==3.5
paho-mqtt==1.6.1
asyncio-mqtt==0.16.1
pyserial-asyncio==0.6

fastapi==0.111.0
uvicorn[standard]==0.30.1

tenacity==8.2.3
python-dotenv==1.0.1
loguru==0.7.2
python-dotenv==1.0.1
openpyxl==3.1.2
